Add linear kernels
