+++ include/flexflow/ops/linear.h
@@ -1,10 +1,7 @@
 #ifndef _FLEXFLOW_LINEAR_H
 #define _FLEXFLOW_LINEAR_H
 
-#include "flexflow/device.h"
-#include "flexflow/fftype.h"
 #include "flexflow/node.h"
-#include "flexflow/op_meta.h"
 #include "flexflow/operator.h"
 #include "flexflow/ops/linear_params.h"
 
@@ -12,23 +9,6 @@ namespace FlexFlow {
 
 class FFModel;
 class Layer;
-
-class LinearMeta : public OpMeta {
-public:
-  LinearMeta(FFHandler handle, int batch_size);
-#if defined(FF_USE_CUDA) || defined(FF_USE_HIP_CUDA)
-  cudnnTensorDescriptor_t outputTensor;
-  cudnnActivationDescriptor_t actiDesc;
-#else
-  miopenTensorDescriptor_t outputTensor;
-  miopenActivationDescriptor_t actiDesc;
-#endif
-  float const *one_ptr;
-  ActiMode activation;
-  bool use_bias;
-  DataType input_type, weight_type, output_type;
-  char op_name[MAX_OPNAME];
-};
 
 class Linear : public Op {
 public:
@@ -67,7 +47,6 @@ public:
                            std::vector<Legion::PhysicalRegion> const &regions,
                            Legion::Context ctx,
                            Legion::Runtime *runtime);
-  static void init_kernel(LinearMeta *m, int batch_size, int channel);
   static void forward_task(Legion::Task const *task,
                            std::vector<Legion::PhysicalRegion> const &regions,
                            Legion::Context ctx,
@@ -76,46 +55,6 @@ public:
                             std::vector<Legion::PhysicalRegion> const &regions,
                             Legion::Context ctx,
                             Legion::Runtime *runtime);
-  static void forward_kernel(LinearMeta const *m,
-                             void const *input_ptr,
-                             void *output_ptr,
-                             void const *filter_ptr,
-                             void const *bias_ptr,
-                             int in_dim,
-                             int out_dim,
-                             int batch_size,
-                             ffStream_t stream);
-  static void forward_kernel_wrapper(LinearMeta const *m,
-                                     void const *input_ptr,
-                                     void *output_ptr,
-                                     void const *filter_ptr,
-                                     void const *bias_ptr,
-                                     int in_dim,
-                                     int out_dim,
-                                     int batch_size);
-  static void backward_kernel(LinearMeta const *m,
-                              void const *input_ptr,
-                              void *input_grad_ptr,
-                              void const *output_ptr,
-                              void *output_grad_ptr,
-                              void const *kernel_ptr,
-                              void *kernel_grad_ptr,
-                              void *bias_ptr,
-                              int in_dim,
-                              int out_dim,
-                              int batch_size,
-                              ffStream_t stream);
-  static void backward_kernel_wrapper(LinearMeta const *m,
-                                      void const *input_ptr,
-                                      void *input_grad_ptr,
-                                      void const *output_ptr,
-                                      void *output_grad_ptr,
-                                      void const *kernel_ptr,
-                                      void *kernel_grad_ptr,
-                                      void *bias_ptr,
-                                      int in_dim,
-                                      int out_dim,
-                                      int batch_size);
   bool measure_operator_cost(Simulator *sim,
                              MachineView const &pc,
                              CostMetrics &cost_metrics) const override;
@@ -162,8 +101,7 @@ private:
       backward_task_with_dim(Legion::Task const *task,
                              std::vector<Legion::PhysicalRegion> const &regions,
                              Legion::Context ctx,
-                             Legion::Runtime *runtime);
-  static bool use_activation(ActiMode mode);
+                             Legion::Runtime *runtime); 
 
   void register_mappings();
   void register_output_mappings();
